rag/requirements.txt
langchain>=0.3
langchain-openai>=0.3
langchain-text-splitters>=0.3
azure-search-documents>=11.6.0b9
azure-identity>=1.19
python-dotenv>=1.0
openai>=1.60
